+++ Sources/Sentry/NSDate+SentryExtras.m
@@ -2,7 +2,8 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-@implementation NSDate (SentryExtras)
+@implementation
+NSDate (SentryExtras)
 
 + (NSDateFormatter *)getIso8601Formatter
 {