+++ Sources/Sentry/SentryTracer.m
@@ -16,6 +16,8 @@
 #import "SentryUIViewControllerPerformanceTracker.h"
 #import <SentryScreenFrames.h>
 
+NS_ASSUME_NONNULL_BEGIN
+
 static const void *spanTimestampObserver = &spanTimestampObserver;
 
 /**
@@ -69,8 +71,8 @@ static BOOL appStartMeasurementRead;
                                        hub:(nullable SentryHub *)hub
                            waitForChildren:(BOOL)waitForChildren
 {
-    if ([super init]) {
-        self.rootSpan = [[SentrySpan alloc] initWithTracer:self context:transactionContext];
+    if (self = [super init]) {
+        self.rootSpan = [[SentrySpan alloc] initWithTransaction:self context:transactionContext];
         self.name = transactionContext.name;
         self.children = [[NSMutableArray alloc] init];
         self.hub = hub;
@@ -119,7 +121,7 @@ static BOOL appStartMeasurementRead;
                                            sampled:_rootSpan.context.sampled];
     context.spanDescription = description;
 
-    SentrySpan *child = [[SentrySpan alloc] initWithTracer:self context:context];
+    SentrySpan *child = [[SentrySpan alloc] initWithTransaction:self context:context];
     @synchronized(self.children) {
         [self.children addObject:child];
     }
@@ -141,17 +143,17 @@ static BOOL appStartMeasurementRead;
     return self.rootSpan.context;
 }
 
-- (NSDate *)timestamp
+- (nullable NSDate *)timestamp
 {
     return self.rootSpan.timestamp;
 }
 
-- (void)setTimestamp:(NSDate *)timestamp
+- (void)setTimestamp:(nullable NSDate *)timestamp
 {
     self.rootSpan.timestamp = timestamp;
 }
 
-- (NSDate *)startTimestamp
+- (nullable NSDate *)startTimestamp
 {
     return self.rootSpan.startTimestamp;
 }
@@ -170,7 +172,7 @@ static BOOL appStartMeasurementRead;
     return _traceState;
 }
 
-- (void)setStartTimestamp:(NSDate *)startTimestamp
+- (void)setStartTimestamp:(nullable NSDate *)startTimestamp
 {
     self.rootSpan.startTimestamp = startTimestamp;
 
@@ -179,7 +181,7 @@ static BOOL appStartMeasurementRead;
 #endif
 }
 
-- (NSDictionary<NSString *, id> *)data
+- (nullable NSDictionary<NSString *, id> *)data
 {
     return self.rootSpan.data;
 }
@@ -194,7 +196,7 @@ static BOOL appStartMeasurementRead;
     return self.rootSpan.isFinished;
 }
 
-- (void)setDataValue:(id)value forKey:(NSString *)key
+- (void)setDataValue:(nullable id)value forKey:(NSString *)key
 {
     [self.rootSpan setDataValue:value forKey:key];
 }
@@ -351,23 +353,28 @@ static BOOL appStartMeasurementRead;
 - (NSArray<SentrySpan *> *)buildAppStartSpans:
     (nullable SentryAppStartMeasurement *)appStartMeasurement
 {
-    if (appStartMeasurement == nil || appStartMeasurement.type == SentryAppStartTypeUnknown) {
+    if (appStartMeasurement == nil) {
+        return @[];
+    }
+
+    NSString *operation;
+    NSString *type;
+
+    switch (appStartMeasurement.type) {
+    case SentryAppStartTypeCold:
+        operation = @"app.start.cold";
+        type = @"Cold Start";
+        break;
+    case SentryAppStartTypeWarm:
+        operation = @"app.start.warm";
+        type = @"Warm Start";
+        break;
+    default:
         return @[];
     }
 
     NSDate *appStartEndTimestamp = [appStartMeasurement.appStartTimestamp
         dateByAddingTimeInterval:appStartMeasurement.duration];
-
-    NSString *operation;
-    NSString *type;
-
-    if (appStartMeasurement.type == SentryAppStartTypeCold) {
-        operation = @"app.start.cold";
-        type = @"Cold Start";
-    } else if (appStartMeasurement.type == SentryAppStartTypeWarm) {
-        operation = @"app.start.warm";
-        type = @"Warm Start";
-    }
 
     SentrySpan *appStartSpan = [self buildSpan:_rootSpan.context.spanId
                                      operation:operation
@@ -456,7 +463,7 @@ static BOOL appStartMeasurementRead;
                                            sampled:_rootSpan.context.sampled];
     context.spanDescription = description;
 
-    return [[SentrySpan alloc] initWithContext:context];
+    return [[SentrySpan alloc] initWithTransaction:self context:context];
 }
 
 - (NSDictionary *)serialize
@@ -483,9 +490,11 @@ static BOOL appStartMeasurementRead;
     if ([span isKindOfClass:[SentryTracer class]]) {
         return span;
     } else if ([span isKindOfClass:[SentrySpan class]]) {
-        return [(SentrySpan *)span tracer];
+        return [(SentrySpan *)span transaction];
     }
     return nil;
 }
 
 @end
+
+NS_ASSUME_NONNULL_END