+++ Tests/SentryTests/Integrations/SentryUIViewControllerSwizziling+Test.h
@@ -4,11 +4,16 @@ NS_ASSUME_NONNULL_BEGIN
 
 #if SENTRY_HAS_UIKIT
 
-@interface SentryUIViewControllerSwizziling (Test)
+@interface
+SentryUIViewControllerSwizziling (Test)
 
-+ (BOOL)shouldSwizzleViewController:(Class)class;
+- (BOOL)shouldSwizzleViewController:(Class)class;
 
-+ (void)swizzleViewControllerSubClass:(Class)class;
+- (void)swizzleViewControllerSubClass:(Class)class;
+
+- (void)swizzleSubclassesOf:(Class)parentClass
+              dispatchQueue:(SentryDispatchQueueWrapper *)dispatchQueue
+               swizzleBlock:(void (^)(Class))block;
 
 @end