+++ Tests/SentryTests/Integrations/SentryUIViewControllerSwizzlingTests.swift
@@ -3,37 +3,44 @@ import XCTest
 #if os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
 
 class SentryUIViewControllerSwizzlingTests: XCTestCase {
+    
+    private class Fixture {
+        var sut: SentryUIViewControllerSwizziling {
+            let options = Options()
+            let imageName = String(
+                cString: class_getImageName(SentryUIViewControllerSwizzlingTests.self)!,
+                encoding: .utf8)! as NSString
+            options.add(inAppInclude: imageName.lastPathComponent)
+            return SentryUIViewControllerSwizziling(options: options, dispatchQueue: TestSentryDispatchQueueWrapper())
+        }
+    }
+    
+    private var fixture: Fixture!
 
     override func setUp() {
         super.setUp()
-        let options = Options()
-        let imageName = String(
-            cString: class_getImageName(SentryUIViewControllerSwizzlingTests.self)!,
-            encoding: .utf8)! as NSString
-        options.add(inAppInclude: imageName.lastPathComponent)
-        SentryUIViewControllerSwizziling.start(with: options, dispatchQueue: TestSentryDispatchQueueWrapper())
+        fixture = Fixture()
     }
     
     override func tearDown() {
         super.tearDown()
-        SentryUIViewControllerSwizziling.start(with: Options(), dispatchQueue: TestSentryDispatchQueueWrapper())
         clearTestState()
     }
 
     func testShouldSwizzle_TestViewController() {
-        let result = SentryUIViewControllerSwizziling.shouldSwizzleViewController(TestViewController.self)
+        let result = fixture.sut.shouldSwizzleViewController(TestViewController.self)
 
         XCTAssertTrue(result)
     }
     
     func testShouldNotSwizzle_NoImageClass() {
-        let result = SentryUIViewControllerSwizziling.shouldSwizzleViewController(UIApplication.self)
+        let result = fixture.sut.shouldSwizzleViewController(UIApplication.self)
 
         XCTAssertFalse(result)
     }
     
     func testShouldNotSwizzle_UIViewController() {
-        let result = SentryUIViewControllerSwizziling.shouldSwizzleViewController(UIViewController.self)
+        let result = fixture.sut.shouldSwizzleViewController(UIViewController.self)
 
         XCTAssertFalse(result)
     }
@@ -47,6 +54,7 @@ class SentryUIViewControllerSwizzlingTests: XCTestCase {
     }
     
     func testViewControllerWithLoadView_TransactionBoundToScope() {
+        fixture.sut.start()
         let controller = ViewWithLoadViewController()
         
         controller.loadView()
@@ -58,7 +66,45 @@ class SentryUIViewControllerSwizzlingTests: XCTestCase {
         let expectedTransactionName = SentryUIViewControllerSanitizer.sanitizeViewControllerName( controller)
         XCTAssertEqual(expectedTransactionName, transactionName)
     }
+    
+    func testSwizzleSubclassesOfParent() {
+        testSwizzleSubclassesOf(Parent.self, expected: [Child1.self, Child2.self, GrandChild1.self, GrandChild2.self])
+    }
+    
+    func testSwizzleSubclassesOfChild1() {
+        testSwizzleSubclassesOf(Child1.self, expected: [GrandChild2.self, GrandChild1.self])
+    }
 
+    func testSwizzleSubclassesOfChild2() {
+        testSwizzleSubclassesOf(Child2.self, expected: [])
+    }
+    
+    private func testSwizzleSubclassesOf(_ type: AnyClass, expected: [AnyClass]) {
+        let expect = expectation(description: "")
+        
+        if expected.isEmpty {
+            expect.isInverted = true
+        } else {
+            expect.expectedFulfillmentCount = expected.count
+        }
+        
+        var actual: [AnyClass] = []
+        fixture.sut.swizzleSubclasses(of: type, dispatchQueue: SentryDispatchQueueWrapper()) { subClass in
+            XCTAssertTrue(Thread.isMainThread, "Block must be executed on the main thread.")
+            actual.append(subClass)
+            expect.fulfill()
+        }
+        
+        wait(for: [expect], timeout: 1)
+        
+        let count = actual.filter { element in
+            return expected.contains { ex in
+                return element == ex
+            }
+        }.count
+        
+        XCTAssertEqual(expected.count, count)
+    }
 }
 
 class ViewWithLoadViewController: UIViewController {
@@ -67,5 +113,11 @@ class ViewWithLoadViewController: UIViewController {
         // empty on purpose
     }
 }
+
+class Parent: UIViewController {}
+class Child1: Parent {}
+class Child2: Parent {}
+class GrandChild1: Child1 {}
+class GrandChild2: Child1 {}
 
 #endif